requirements.txt
amqp==5.1.1
asgiref==3.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.7
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
kombu==5.2.4
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
ujson==5.7.0
vine==5.0.0
wcwidth==0.2.6
